backend/requirements.txt
# asgiref==3.8.1
# bleach==6.2.0
# channels==3.0.5
# django-prometheus==2.2.0
# certifi==2024.8.30
# cffi==1.17.1
# charset-normalizer==3.4.0
# cryptography==44.0.0
# defusedxml==0.8.0rc2
# djangorestframework-simplejwt==5.3.1
# Django==5.1.4
# PyJWT==2.10.1
# django-cors-headers==4.6.0
# django-extensions==3.2.3
# django-friendship==1.9.6
# djangorestframework==3.15.2
# idna==3.10
# oauthlib==3.2.2
# psycopg==3.2.3
# psycopg2-binary==2.9.10
# pycparser==2.22
# python-dotenv==1.0.1
# python3-openid==3.2.0
# requests==2.32.3
# requests-oauthlib==2.0.0
# sqlparse==0.5.2
# typing_extensions==4.12.2
# urllib3==2.2.3
# webencodings==0.5.1
# psycopg>=3.1.10
# psycopg2-binary>=2.9.7


asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
channels==3.0.5
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
defusedxml==0.8.0rc2
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-friendship==1.9.6
django-prometheus==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.2.2
pillow==11.0.0
prometheus_client==0.21.1
psycopg==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.6.0
sqlparse==0.5.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
zope.interface==7.2
